RichTextLabel: add furigana tag
Previous to this commit there was no easy way to add a text on
top of each other in a RichTextLabel. This is quite essential for
Japanese apps and games that rely on difficult to read Kanji and
its Furigana on top of them.

So the [furigana] tag is added as one of the possible bbcodes to
call. This will make some text be displayed on top of another.

To use it just add the text "[furigana=top_text]text[/furigana]"
into a RichTextLabel and it will display the text with top_text on
top.

Since it is common to have the furigana text be a smaller size than
the rest of the text, a furigana font has been added to allow users
change the font for this text.